requirements.txt
anyio==4.12.0
certifi==2025.11.12
exceptiongroup==1.3.1
greenlet==3.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
packaging==25.0
playwright==1.57.0
pluggy==1.6.0
pyee==13.0.0
Pygments==2.19.2
pytest==9.0.2
python-dotenv==1.2.1
python-telegram-bot==22.5
tinydb==4.8.2
tomli==2.3.0
typing_extensions==4.15.0
